Move commands to config
